backend/requirements.txt
fastapi
uvicorn
gunicorn
python-multipart
pydantic
python-dotenv
# Core dependencies from original project

google-genai
litellm
python-docx
python-pptx
pandas
fpdf
watchdog
sqlalchemy
psycopg2-binary
alembic
slowapi
passlib[bcrypt]
redis>=4.0
aioredis>=2.0
python-jose[cryptography]
email-validator
httpx
webauthn
cbor2

# Stripe for payment processing
stripe

